Move enums WheelchairAccessibleStatus and BikesAllowedStatus to packages gtfs
