+++ domain/src/test/java/org/mobilitydata/gtfsvalidator/domain/entity/gtfs/BikesAllowedStatusTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.mobilitydata.gtfsvalidator.domain.entity;
+package org.mobilitydata.gtfsvalidator.domain.entity.gtfs;
 
 import org.junit.jupiter.api.Test;
