+++ domain/src/test/java/org/mobilitydata/gtfsvalidator/domain/entity/gtfs/WheelchairAccessibleStatusTest.java
@@ -1,4 +1,4 @@
-package org.mobilitydata.gtfsvalidator.domain.entity;
+package org.mobilitydata.gtfsvalidator.domain.entity.gtfs;
 
 import org.junit.jupiter.api.Test;
